refactor: move ome-zarr code into reusable module
